refactor(session): split the vocabulary into ten files
