+++ packages/AGENTS.md
@@ -68,3 +68,8 @@ for. A package that holds both a wire vocabulary and a Node flow gives the
 vocabulary a subpath of its own (`@sidecar/calendar/vocabulary`,
 `@sidecar/account/snapshot`, `@sidecar/superset/sign-in-stage`), or the renderer
 bundle fails to resolve `node:http` behind a string constant it wanted to draw.
+
+A barrel over modules that are all one vocabulary is written as `export *` per
+module (`@sidecar/session`), because a hand-listed re-export of a package whose
+every name is public is a second list to forget a name in; a package whose
+modules are not all public keeps the explicit list, which is what says so.
