+++ packages/session/src/normalize.test.ts
@@ -1,26 +1,8 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import {
-  normalizeSession,
-  SESSION_STATUS,
-  type Session,
-  type SessionStatus,
-  sessionChangeNumber,
-} from "@sidecar/session";
+import { normalizeSession, SESSION_STATUS } from "@sidecar/session";
 
 const TEST_NOW = Date.parse("2026-08-16T12:00:00.000Z");
-const _DAY_MS = 24 * 60 * 60 * 1000;
-
-function _session(
-  providerSessionId: string,
-  status: SessionStatus,
-  lastActivityAt: number,
-): Session {
-  return normalizeSession(
-    { id: "codex", displayName: "Codex" },
-    { providerSessionId, title: "Implement the shared session core", status, lastActivityAt },
-  );
-}
 
 test("the grouping manager's mark leads the row and the press follows it", () => {
   const normalized = normalizeSession(
@@ -155,27 +137,6 @@ test("the Claude app's own scheme is an address a row may open", () => {
 
   assert.equal(normalized.detail.link, "claude://claude.ai/epitaxy/local_desk-1");
   assert.equal(normalized.applications[0]?.link, "claude://claude.ai/epitaxy/local_desk-1");
-});
-
-test("reads the pull request's number off every host's address shape", () => {
-  assert.equal(sessionChangeNumber("https://github.com/reviewstage/luke/pull/245"), 245);
-  assert.equal(
-    sessionChangeNumber("https://gitlab.com/reviewstage/group/sidecar-web/-/merge_requests/3"),
-    3,
-  );
-  assert.equal(
-    sessionChangeNumber("https://bitbucket.org/reviewstage/sidecar-native/pull-requests/9"),
-    9,
-  );
-});
-
-test("an address whose tail names no number yields none rather than a guess", () => {
-  assert.equal(sessionChangeNumber("https://github.com/reviewstage/luke/pulls"), undefined);
-  assert.equal(
-    sessionChangeNumber("https://github.com/reviewstage/luke/pull/245/files"),
-    undefined,
-  );
-  assert.equal(sessionChangeNumber("not an address"), undefined);
 });
 
 test("keeps a sound diff summary and drops a suspect or empty one whole", () => {
