--- packages/session/src/providers.ts
@@ -1,848 +1,0 @@
-import {
-  ACT_RESULT_STATUS,
-  type ActResult,
-  text,
-  type UnknownActResult,
-  type UnparsedWireValue,
-} from "@sidecar/wire";
-import {
-  type ProviderSessionObservation,
-  SESSION_APPLICATION_ID,
-  type SessionControl,
-  type SessionProvider,
-} from "./session.js";
-
-export const PROVIDER_LOCATION_KIND = {
-  LOCAL: "local",
-  CLOUD: "cloud",
-  LOCAL_AND_CLOUD: "local-and-cloud",
-} as const;
-
-export type ProviderLocationKind =
-  (typeof PROVIDER_LOCATION_KIND)[keyof typeof PROVIDER_LOCATION_KIND];
-
-/**
- * Stable provider identifiers shared by adapters, the registry, and the UI.
- * They key provider-specific facts without a renderer importing adapter code
- * or matching on a display name.
- */
-export const PROVIDER_ID = {
-  CLAUDE_CODE: "claude-code",
-  CODEX: "codex",
-  CONDUCTOR: "conductor",
-  OMP: "omp",
-} as const;
-
-export type ProviderId = (typeof PROVIDER_ID)[keyof typeof PROVIDER_ID];
-
-export interface ProviderIdentity {
-  readonly id: ProviderId;
-  readonly displayName: string;
-  readonly location: ProviderLocationKind;
-}
-
-/**
- * Shared provider identity only. Capabilities, credentials, hooks, fixtures,
- * and presentation stay in the packages that own those decisions.
- */
-export const PROVIDER_IDENTITY_BY_ID = {
-  [PROVIDER_ID.CLAUDE_CODE]: {
-    id: PROVIDER_ID.CLAUDE_CODE,
-    displayName: "Claude Code",
-    location: PROVIDER_LOCATION_KIND.LOCAL,
-  },
-  [PROVIDER_ID.CODEX]: {
-    id: PROVIDER_ID.CODEX,
-    displayName: "Codex",
-    location: PROVIDER_LOCATION_KIND.LOCAL_AND_CLOUD,
-  },
-  [PROVIDER_ID.CONDUCTOR]: {
-    id: PROVIDER_ID.CONDUCTOR,
-    displayName: "Conductor",
-    location: PROVIDER_LOCATION_KIND.CLOUD,
-  },
-  [PROVIDER_ID.OMP]: {
-    id: PROVIDER_ID.OMP,
-    displayName: "OMP",
-    location: PROVIDER_LOCATION_KIND.LOCAL,
-  },
-} as const satisfies Readonly<Record<ProviderId, ProviderIdentity>>;
-
-/**
- * The provider id the local Conductor workspace creator answers to. It names
- * no observed session provider — `isProviderId` stays false for it — because a
- * local Conductor chat is already observed by the agent that runs it. The id
- * exists only so the repositories Conductor holds can offer their creation
- * projects apart from the cloud Conductor adapter, which owns
- * `PROVIDER_ID.CONDUCTOR`: a creation request resolves to exactly one adapter,
- * so a local repository and a cloud project must route under ids of their own,
- * even as both wear Conductor's name and mark.
- */
-export const CONDUCTOR_LOCAL_WORKSPACE_PROVIDER_ID = "conductor-local";
-
-export const SUPERSET_WORKSPACE_PROVIDER_ID = SESSION_APPLICATION_ID.SUPERSET;
-
-/**
- * Every provider that can offer a workspace through the desktop app: the
- * observed session providers, Superset's own workspace provider, and local
- * Conductor's — the last two name no observed session provider, so they are
- * added beside `ProviderId` rather than found within it.
- */
-export type WorkspaceProviderId =
-  | ProviderId
-  | typeof SUPERSET_WORKSPACE_PROVIDER_ID
-  | typeof CONDUCTOR_LOCAL_WORKSPACE_PROVIDER_ID;
-
-/**
- * The order any list of providers reads in. It is the registry's own order
- * rather than one derived from live sessions, so a list of agents does not
- * reshuffle as their sessions come and go.
- */
-export const PROVIDER_ID_LIST: readonly ProviderId[] = Object.values(PROVIDER_IDENTITY_BY_ID).map(
-  (identity) => identity.id,
-);
-
-/**
- * Agents Luke draws only inside a hosting app's workspaces — the agents
- * Conductor and Superset run beside Claude Code and Codex. Identity alone:
- * the id keys the agent's own mark and filter chip, and nothing else exists
- * behind it — no adapter, no files, no credential — so the hosting provider
- * stays the thing observed, credentialed, and written through.
- */
-export const HOSTED_AGENT_ID = {
-  COPILOT: "copilot",
-  CURSOR: "cursor",
-  GEMINI_CLI: "gemini-cli",
-  GROK_BUILD: "grok-build",
-  OPENCODE: "opencode",
-} as const;
-
-export type HostedAgentId = (typeof HOSTED_AGENT_ID)[keyof typeof HOSTED_AGENT_ID];
-
-/** The registry's own order, for the same reason `PROVIDER_ID_LIST` keeps one. */
-export const HOSTED_AGENT_ID_LIST: readonly HostedAgentId[] = Object.values(HOSTED_AGENT_ID);
-
-const HOSTED_AGENT_IDS: ReadonlySet<string> = new Set(HOSTED_AGENT_ID_LIST);
-
-/** Whether this build draws the hosted agent an observation names. */
-export function isHostedAgentId(value: string): value is HostedAgentId {
-  return HOSTED_AGENT_IDS.has(value);
-}
-
-const PROVIDER_IDS: ReadonlySet<string> = new Set(PROVIDER_ID_LIST);
-
-/** Whether this build knows the provider an observation names. */
-export function isProviderId(value: string): value is ProviderId {
-  return PROVIDER_IDS.has(value);
-}
-
-export function isWorkspaceProviderId(value: string): value is WorkspaceProviderId {
-  return (
-    isProviderId(value) ||
-    value === SUPERSET_WORKSPACE_PROVIDER_ID ||
-    value === CONDUCTOR_LOCAL_WORKSPACE_PROVIDER_ID
-  );
-}
-
-/**
- * A provider adapter has no dependency on Electron, a renderer, or live UI
- * state. Every adapter answers every act, because unsupported is already an
- * answer rather than a failure: one whose provider documents no way to do a
- * thing inherits that answer from `SessionProviderAdapterBase` and says so,
- * where a missing method would leave each caller asking whether the question
- * could be put at all. Overriding one is what taking on its constraint means.
- */
-export interface SessionProviderAdapter {
-  readonly provider: SessionProvider;
-  observe(): Promise<readonly ProviderSessionObservation[]>;
-
-  /**
-   * Runs a control against a session the adapter has already observed.
-   * Adapters must reject any request whose control that session's latest
-   * observation did not advertise.
-   */
-  executeControl(request: ProviderControlRequest): Promise<ProviderControlResult>;
-
-  /**
-   * Hands a message to an already-observed session through the provider's own
-   * documented endpoint. It is one of the three places an adapter may change
-   * provider state, and only ever with text a user chose to send: adapters
-   * must refuse any session that did not advertise `canReceiveMessage` on its
-   * latest observation, and nothing that decides on the user's behalf may
-   * reach it.
-   */
-  sendMessage(message: ProviderSessionMessage): Promise<ProviderMessageResult>;
-
-  /** The projects the latest observation pass reported, or none. */
-  workspaceProjects(): readonly WorkspaceProject[];
-
-  /**
-   * Creates a workspace in a project the latest observation pass reported. The
-   * same rules bind it that bind a message: it acts only on what a user just
-   * asked for, through the provider's own documented endpoint, and nothing
-   * that decides on the user's behalf may reach it.
-   */
-  createWorkspace(request: ProviderWorkspaceRequest): Promise<ProviderWorkspaceResult>;
-
-  /**
-   * Starts another agent in the workspace an observed session already runs in,
-   * under the same rules and one more: the agent must be one of the kinds that
-   * session's latest observation listed.
-   */
-  spawnWorkspaceAgent(request: ProviderWorkspaceAgentRequest): Promise<ProviderWorkspaceResult>;
-
-  /**
-   * Renames the workspace an observed session already runs in, under the same
-   * rules and one more: the session's latest observation must have advertised
-   * a rename target, so a rename only ever lands on a workspace its provider
-   * documents renaming.
-   */
-  renameWorkspace(request: ProviderWorkspaceRenameRequest): Promise<ProviderActResult>;
-
-  /**
-   * Renames an observed session itself — the chat, where `renameWorkspace`
-   * renames the workspace around it — only for a session whose latest
-   * observation advertised `canRename`.
-   */
-  renameSession(request: ProviderSessionRenameRequest): Promise<ProviderActResult>;
-
-  /**
-   * Renders one observed session's own transcript, read from the provider's
-   * file on this machine, into a bounded conversation kept nowhere. The read
-   * performs nothing and reaches no provider; an adapter whose stored shape
-   * this build cannot render faithfully reports nothing rather than guessing.
-   */
-  readTranscript(providerSessionId: string): Promise<ProviderTranscriptResult>;
-
-  /**
-   * Renders what one observed session's transcript has gained since the
-   * cursor a previous read handed back, under the same bounds and the same
-   * rules as `readTranscript`: read from the provider's own record, kept
-   * nowhere, performing nothing. Without a cursor, or with one the record no
-   * longer reaches, it renders the transcript's recent tail and says so
-   * through `truncated`. The cursor is opaque to callers and meaningful only
-   * to the adapter that minted it.
-   */
-  readTranscriptSince(
-    providerSessionId: string,
-    cursor?: string,
-  ): Promise<ProviderTranscriptSinceResult>;
-
-  /**
-   * Reads one observed session's conversation from the provider's documented
-   * transcript-read endpoint, as bounded pages of attributed messages kept
-   * nowhere. It exists only for a caller a user just opened a conversation
-   * screen for: it never rides an observation pass, it can express nothing
-   * but a read, and a message whose author the stored shape does not name is
-   * dropped rather than guessed at. An adapter whose provider documents no
-   * such read inherits the unsupported answer.
-   */
-  readConversation(request: ProviderConversationRequest): Promise<ProviderConversationResult>;
-}
-
-/**
- * What became of a write the user asked for. Every adapter capability answers
- * with the same three: accepted, rejected with a reason the user can act on,
- * or unsupported — the adapter has no documented way to do this, which is an
- * answer rather than a failure. One status set, because two identical triples
- * would be an API break the moment they diverged.
- */
-export type ProviderActResult = ActResult;
-
-/** A provider-local request for a control that was previously exposed by observation. */
-export interface ProviderControlRequest {
-  providerSessionId: string;
-  control: SessionControl;
-}
-
-/**
- * What became of a control. Providers must report unsupported or rejected
- * controls explicitly; the core deliberately provides no fallback path such
- * as terminal input injection.
- */
-export type ProviderControlResult = ProviderActResult;
-
-/** A user-authored message for one session the adapter has already observed. */
-export interface ProviderSessionMessage {
-  providerSessionId: string;
-  text: string;
-}
-
-/**
- * What became of a send. A rejection carries a reason the user can act on,
- * never the message itself; unsupported means the adapter has no documented
- * way to message this session, which is an answer rather than a failure.
- */
-export type ProviderMessageResult = ProviderActResult;
-
-export type ProviderTranscriptResult =
-  | { status: typeof ACT_RESULT_STATUS.ACCEPTED; transcript: string }
-  | { status: typeof ACT_RESULT_STATUS.REJECTED; reason: string }
-  | { status: typeof ACT_RESULT_STATUS.UNSUPPORTED; reason: string };
-
-/**
- * One incremental transcript reading: the lines gained since the cursor the
- * caller held, the cursor to continue from, and whether the read fell short
- * of everything gained. An empty `text` is an honest "nothing new". The
- * cursor is absent only when the provider handed back no position to resume
- * from, so the next read begins as this one did.
- */
-export interface ProviderTranscriptSinceReading {
-  text: string;
-  cursor?: string;
-  truncated: boolean;
-}
-
-export type ProviderTranscriptSinceResult =
-  | ({ status: typeof ACT_RESULT_STATUS.ACCEPTED } & ProviderTranscriptSinceReading)
-  | { status: typeof ACT_RESULT_STATUS.REJECTED; reason: string }
-  | { status: typeof ACT_RESULT_STATUS.UNSUPPORTED; reason: string };
-
-/**
- * Who wrote one message of a conversation reading. Only the two voices a chat
- * screen draws exist here: the developer's own sends and the agent's own
- * words. Everything else a provider stores beside them — tool calls, tool
- * output, lifecycle events, harness chatter — has no author a bubble can
- * honestly wear, so it never becomes a `ProviderConversationMessage` at all.
- */
-export const CONVERSATION_MESSAGE_AUTHOR = {
-  USER: "user",
-  AGENT: "agent",
-} as const;
-
-export type ConversationMessageAuthor =
-  (typeof CONVERSATION_MESSAGE_AUTHOR)[keyof typeof CONVERSATION_MESSAGE_AUTHOR];
-
-/**
- * One attributed message of an observed session's conversation, exactly as
- * the provider stored it: the words are never truncated — a cut message says
- * something its author did not — and the bounds live on the page instead.
- */
-export interface ProviderConversationMessage {
-  /** The provider's own id for this message. */
-  id: string;
-  author: ConversationMessageAuthor;
-  text: string;
-  /** Unix ms the provider recorded the message at, when it reported one. */
-  receivedAt?: number;
-}
-
-/**
- * A user-asked read of one observed session's conversation, positioned the
- * way a chat screen reads: opened with neither cursor, it answers the latest
- * page; handed `beforeOffset`, it answers the page of older history ending
- * there; handed `afterMessageId`, it answers only what is newer. The two
- * cursors are different asks — a scroll up and a poll — so a request naming
- * both is refused rather than guessed at.
- */
-export interface ProviderConversationRequest {
-  providerSessionId: string;
-  /**
-   * The last provider message id an earlier page answered with, so a polling
-   * screen reads only what is newer.
-   */
-  afterMessageId?: string;
-  /**
-   * The stored-transcript offset an earlier page said it began at
-   * (`firstOffset`), so a scroll to the top reads the history just before
-   * what the screen already holds.
-   */
-  beforeOffset?: number;
-}
-
-/**
- * What a conversation read answered with: one bounded page of attributed
- * messages and the positions to continue from. `lastMessageId` names the
- * newest stored message the page consumed — attributed or not — so a poll
- * resumes where this read stopped; it is absent on an older-history page,
- * which must never move the poll cursor backward. `firstOffset` names the
- * stored offset the page began at and `hasOlder` whether history stands
- * before it, so a scroll to the top can keep reading; both are absent on a
- * poll, which never looks backward. `hasMore` says newer messages remain
- * beyond a bounded poll page.
- */
-export type ProviderConversationResult =
-  | {
-      status: typeof ACT_RESULT_STATUS.ACCEPTED;
-      messages: readonly ProviderConversationMessage[];
-      lastMessageId?: string;
-      hasMore: boolean;
-      firstOffset?: number;
-      hasOlder?: boolean;
-    }
-  | { status: typeof ACT_RESULT_STATUS.REJECTED; reason: string }
-  | { status: typeof ACT_RESULT_STATUS.UNSUPPORTED; reason: string };
-
-export async function providerTranscriptResult(
-  rendering: Promise<string | undefined>,
-): Promise<ProviderTranscriptResult> {
-  const transcript = await rendering;
-  return transcript
-    ? { status: ACT_RESULT_STATUS.ACCEPTED, transcript }
-    : {
-        status: ACT_RESULT_STATUS.REJECTED,
-        reason: "That session's transcript could not be found.",
-      };
-}
-
-export async function providerTranscriptSinceResult(
-  reading: Promise<ProviderTranscriptSinceReading | undefined>,
-): Promise<ProviderTranscriptSinceResult> {
-  const read = await reading;
-  return read
-    ? { status: ACT_RESULT_STATUS.ACCEPTED, ...read }
-    : {
-        status: ACT_RESULT_STATUS.REJECTED,
-        reason: "That session's transcript could not be found.",
-      };
-}
-
-/**
- * Whether a new workspace in a project carries an opening task — the
- * developer's own words for what its agent should start on. A provider whose
- * creation endpoint requires a prompt cannot make an idle workspace, and one
- * that documents no way to hand a task at creation cannot take one; each
- * project says which it is, so an ask can be validated before a request
- * exists.
- */
-export const WORKSPACE_TASK_SUPPORT = {
-  NONE: "none",
-  OPTIONAL: "optional",
-  REQUIRED: "required",
-} as const;
-
-export type WorkspaceTaskSupport =
-  (typeof WORKSPACE_TASK_SUPPORT)[keyof typeof WORKSPACE_TASK_SUPPORT];
-
-const WORKSPACE_TASK_SUPPORT_LIST: readonly WorkspaceTaskSupport[] =
-  Object.values(WORKSPACE_TASK_SUPPORT);
-
-/**
- * One place a provider will create a workspace: a project it reported on the
- * latest observation pass. A request can only name one of these, so the set of
- * places a workspace can be asked for is the set the provider itself listed —
- * never a repository URL or path composed on this side.
- */
-export interface WorkspaceProject {
-  /** The provider-owned identifier a creation request names the project by. */
-  providerProjectId: string;
-  /** The repository label the project is named by out loud and on screen. */
-  repository: string;
-  /** Whether a new workspace here takes — or needs — an opening task. */
-  taskSupport: WorkspaceTaskSupport;
-  /** The provider-owned host or execution target that owns this project. */
-  providerTargetId?: string;
-  /** The bounded label a person uses to distinguish that target. */
-  targetName?: string;
-  /** Agent kinds the provider currently permits for a new workspace here. */
-  spawnableAgents?: readonly string[];
-  /** The saved agent kind used when a creation ask names none. */
-  defaultAgent?: string;
-  /**
-   * The provider names a workspace here itself and takes none from the ask,
-   * so an ask carrying one is refused rather than half honoured.
-   */
-  namesItself?: boolean;
-}
-
-/** A workspace project as the app reports it, stamped with who offered it. */
-export interface ObservedWorkspaceProject extends WorkspaceProject {
-  providerId: string;
-  providerName: string;
-}
-
-/** The identity a saved default uses, including a host when one owns it. */
-export function workspaceProjectSelectionId(
-  project: Pick<WorkspaceProject, "providerProjectId" | "providerTargetId">,
-): string {
-  return project.providerTargetId
-    ? JSON.stringify([project.providerProjectId, project.providerTargetId])
-    : project.providerProjectId;
-}
-
-/**
- * The providers whose stored default names no project they currently offer —
- * a choice that steers nothing, because every path that reads it matches
- * against the offered set. Only providers present in `projects` are judged: a
- * provider offering nothing is observing nothing, and a default must not be
- * discarded on that silence.
- */
-export function staleWorkspaceProjectDefaults(
-  projects: readonly ObservedWorkspaceProject[],
-  defaults: Readonly<Partial<Record<string, string>>> | undefined,
-): readonly string[] {
-  if (!defaults) return [];
-  const offered = new Map<string, Set<string>>();
-  for (const project of projects) {
-    const selections = offered.get(project.providerId) ?? new Set<string>();
-    selections.add(workspaceProjectSelectionId(project));
-    offered.set(project.providerId, selections);
-  }
-  return [...offered.entries()]
-    .filter(([providerId, selections]) => {
-      const stored = defaults[providerId];
-      return stored !== undefined && !selections.has(stored);
-    })
-    .map(([providerId]) => providerId);
-}
-
-/** A workspace name reads in one breath; anything longer is a different ask. */
-export const maximumWorkspaceNameLength = 80;
-
-/** How many projects the app will offer workspace creation in at once. */
-export const maximumObservedWorkspaceProjects = 20;
-
-/**
- * The name a new workspace was asked for under, or nothing. Refused rather
- * than cut when it runs long, the same posture as a message: a truncated name
- * says something its author did not.
- */
-export function workspaceNameText(value: UnparsedWireValue): string | undefined {
-  const normalized = text(value);
-  if (!normalized || normalized.length > maximumWorkspaceNameLength) return undefined;
-  return normalized;
-}
-
-/**
- * Bounds, deduplicates, and alphabetizes the projects adapters offered, so
- * the surface and the conversation are handed the same capped, display-safe
- * list. Ordered by the repository label a person scans for — not by which
- * adapter answered first — and capped after the sort, so a list too long to
- * keep whole loses its alphabetical tail rather than an arbitrary provider.
- */
-export function normalizeObservedWorkspaceProjects(
-  projects: readonly ObservedWorkspaceProject[],
-  preferredSelections?: Readonly<Partial<Record<string, string>>>,
-): readonly ObservedWorkspaceProject[] {
-  const seen = new Map<string, Map<string, Set<string>>>();
-  const normalized: ObservedWorkspaceProject[] = [];
-  for (const project of projects) {
-    const providerId = project.providerId.trim();
-    const providerProjectId = project.providerProjectId.trim();
-    const providerTargetId = project.providerTargetId?.trim() ?? "";
-    const repository = project.repository.trim().slice(0, maximumWorkspaceNameLength);
-    if (!providerId || !providerProjectId || !repository) continue;
-    const byProvider = seen.get(providerId) ?? new Map<string, Set<string>>();
-    const byProject = byProvider.get(providerProjectId) ?? new Set<string>();
-    if (byProject.has(providerTargetId)) continue;
-    byProject.add(providerTargetId);
-    byProvider.set(providerProjectId, byProject);
-    seen.set(providerId, byProvider);
-    const normalizedProject: ObservedWorkspaceProject = {
-      providerId,
-      providerName: project.providerName.trim() || providerId,
-      providerProjectId,
-      repository,
-      // A support level this build does not know is read as none, so an ask
-      // is refused rather than guessed at.
-      taskSupport: WORKSPACE_TASK_SUPPORT_LIST.includes(project.taskSupport)
-        ? project.taskSupport
-        : WORKSPACE_TASK_SUPPORT.NONE,
-    };
-    if (providerTargetId) normalizedProject.providerTargetId = providerTargetId;
-    const targetName = project.targetName?.trim();
-    if (targetName) {
-      normalizedProject.targetName = targetName.slice(0, maximumWorkspaceNameLength);
-    }
-    if (project.spawnableAgents) {
-      normalizedProject.spawnableAgents = [
-        ...new Set(project.spawnableAgents.map((agent) => agent.trim())),
-      ]
-        .filter(Boolean)
-        .slice(0, 20);
-    }
-    const defaultAgent = project.defaultAgent?.trim();
-    if (defaultAgent) normalizedProject.defaultAgent = defaultAgent;
-    if (project.namesItself === true) normalizedProject.namesItself = true;
-    normalized.push(normalizedProject);
-  }
-  const sorted = normalized.sort(compareWorkspaceProjects);
-  const preferred = sorted.filter(
-    (project) => preferredSelections?.[project.providerId] === workspaceProjectSelectionId(project),
-  );
-  const ordinary = sorted.filter(
-    (project) => preferredSelections?.[project.providerId] !== workspaceProjectSelectionId(project),
-  );
-  return [...preferred, ...ordinary]
-    .slice(0, maximumObservedWorkspaceProjects)
-    .sort(compareWorkspaceProjects);
-}
-
-function compareWorkspaceProjects(
-  left: ObservedWorkspaceProject,
-  right: ObservedWorkspaceProject,
-): number {
-  return (
-    compareRepositoryLabels(left.repository, right.repository) ||
-    // Two providers can offer one repository label; the provider and then the
-    // id keep the order deterministic rather than arrival-dependent.
-    left.providerName.localeCompare(right.providerName) ||
-    left.providerProjectId.localeCompare(right.providerProjectId)
-  );
-}
-
-/** Alphabetical the way a person reads labels: case-blind, digits as numbers. */
-function compareRepositoryLabels(left: string, right: string): number {
-  return left.localeCompare(right, undefined, { sensitivity: "base", numeric: true });
-}
-
-/**
- * One model an agent runs: the id its provider's endpoints take, and the name
- * a person reads. The label stands alone — no vendor beside it — because a
- * model's own name already says whose it is.
- */
-export interface WorkspaceAgentModel {
-  id: string;
-  label: string;
-}
-
-/**
- * One agent kind a provider's creation endpoints take, with the models it
- * runs and the effort levels it thinks at — all exactly as the provider
- * documents them. An app declares these as build-fixed tables, the way it
- * fixes any other documented value set, so a choice is only ever offered from
- * what the build knows the provider takes. An agent that takes no effort
- * levels documents an empty list, and is simply never offered one.
- */
-export interface WorkspaceAgentModels {
-  agent: string;
-  models: readonly WorkspaceAgentModel[];
-  efforts: readonly string[];
-}
-
-/**
- * The agent kind, model, and optionally effort a user chose for new
- * workspaces, as one value on purpose: a model id or an effort level only
- * means anything beside the agent that runs it, and fields stored apart could
- * recombine into a pairing no table ever listed. Effort is optional inside
- * the pair — absent, the provider's own default effort stands.
- */
-export interface WorkspaceAgentSelection {
-  /** The agent kind, exactly as the provider's documented set names it. */
-  agent: string;
-  /** The model id, exactly as documented for that agent. */
-  model: string;
-  /** The effort level, exactly as documented for that agent, when chosen. */
-  effort?: string;
-}
-
-/**
- * The agent kind a user chose for new workspaces on a provider whose agent
- * kinds are observed presets rather than a build-fixed table, and which
- * documents no model choice (Superset today). The `never` fields are the
- * boundary stated as a type: no model or effort can ride beside a kind the
- * provider's endpoints take alone.
- */
-export interface WorkspaceAgentKindSelection {
-  agent: string;
-  model?: never;
-  effort?: never;
-}
-
-/**
- * The chosen defaults for new workspaces, one entry per workspace provider
- * that documents an agent choice at all: a provider with a build-fixed models
- * table holds a full selection, and Superset holds the kind alone. Local
- * Conductor's creation link documents no agent, model, or name, so it can
- * hold no entry.
- */
-export type WorkspaceAgentDefaults = Readonly<
-  Partial<Record<ProviderId, WorkspaceAgentSelection>> &
-    Partial<Record<typeof SUPERSET_WORKSPACE_PROVIDER_ID, WorkspaceAgentKindSelection>>
->;
-
-/** A user-asked request for a new workspace in one reported project. */
-export interface ProviderWorkspaceRequest {
-  providerProjectId: string;
-  providerTargetId?: string;
-  agent?: string;
-  /** The name the user chose, when they chose one; the provider names it otherwise. */
-  name?: string;
-  /**
-   * The opening task for the workspace's agent, in the user's own words —
-   * present only when the user gave one, and only for a project whose
-   * `taskSupport` takes it. It is the same class of content as a message to
-   * an existing session, and it travels under the same rules.
-   */
-  task?: string;
-  /**
-   * The agent kind and model the user chose for new workspaces, present only
-   * when they chose one and only from the build's documented table for this
-   * provider. Absent, the provider's own defaults decide — the adapter sends
-   * nothing at all rather than a guess.
-   */
-  agentSelection?: WorkspaceAgentSelection;
-}
-
-/**
- * What became of a creation ask — the same three answers a message gets, for
- * the same reasons: a rejection carries a reason the user can act on, and
- * unsupported means the provider documents no way to create one here. An
- * acceptance may also carry the id of the session the creation response
- * named — an identifier only, never an address — so the surface can open the
- * new workspace once an observation pass reports it under that id. A provider
- * whose response names no session simply omits it, and the workspace stands
- * unopened rather than guessed at.
- */
-export type ProviderWorkspaceResult =
-  | {
-      status: typeof ACT_RESULT_STATUS.ACCEPTED;
-      /** The created session's id, exactly as the provider's response named it. */
-      providerSessionId?: string;
-      /** Creation landed, but a non-essential follow-up such as opening failed. */
-      warning?: string;
-    }
-  | { status: typeof ACT_RESULT_STATUS.REJECTED; reason: string }
-  | { status: typeof ACT_RESULT_STATUS.UNSUPPORTED; reason: string }
-  /** The create was handed to the machine and its answer lost: it may have happened, so it is never retried. */
-  | UnknownActResult;
-
-/**
- * Thrown by an open port when the address was handed to the process that
- * opens it and that process went away before answering: the system may have
- * opened it. An adapter that creates through a deep link reads this as an
- * unknown outcome, never a refusal, so the journal can neither call the
- * create failed nor repeat it.
- */
-export class ExternalOpenAnswerLostError extends Error {}
-
-/**
- * A user-asked request for another agent in the workspace an observed session
- * already runs in. The session names the workspace; the agent must be one that
- * session's latest observation listed as spawnable.
- */
-export interface ProviderWorkspaceAgentRequest {
-  providerSessionId: string;
-  /** The kind of agent, exactly as the observation listed it. */
-  agent: string;
-  /** The name the user chose, when they chose one. */
-  name?: string;
-  /** The new agent's opening task, in the user's own words, when they gave one. */
-  task?: string;
-  /**
-   * The model the user's stored choice names for exactly this agent kind,
-   * present only when the kinds match and the pairing is in the build's
-   * documented table. The asked-for agent always wins over a stored pairing:
-   * a preference rides along with the user's ask, never against it.
-   */
-  model?: string;
-  /** The effort level riding with that model, under exactly the same rules. */
-  effort?: string;
-}
-
-/**
- * A user-asked request to rename the workspace an observed session already
- * runs in. The session names the workspace — the adapter resolves the target
- * from its own latest observation, never from the request — and the name is
- * the user's own choice, bounded like the one a creation carries.
- */
-export interface ProviderWorkspaceRenameRequest {
-  providerSessionId: string;
-  /** The new name, exactly as the user chose it. */
-  name: string;
-}
-
-/**
- * A user-asked request to rename one observed session itself — the chat,
- * where `ProviderWorkspaceRenameRequest` renames the workspace around it —
- * under the same rules.
- */
-export interface ProviderSessionRenameRequest {
-  providerSessionId: string;
-  /** The new name, exactly as the user chose it. */
-  name: string;
-}
-
-/**
- * The explicit answers an adapter gives for acts its provider does not
- * document: unsupported for every write, no projects, and no transcript.
- * Concrete adapters override only the acts their provider routes, and an
- * override takes on that act's constraint above along with it.
- */
-export abstract class SessionProviderAdapterBase implements SessionProviderAdapter {
-  abstract readonly provider: SessionProvider;
-  abstract observe(): Promise<readonly ProviderSessionObservation[]>;
-
-  async executeControl(_request: ProviderControlRequest): Promise<ProviderControlResult> {
-    return { status: ACT_RESULT_STATUS.UNSUPPORTED, reason: "This provider has no such control." };
-  }
-
-  async sendMessage(_message: ProviderSessionMessage): Promise<ProviderMessageResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider does not take messages.",
-    };
-  }
-
-  workspaceProjects(): readonly WorkspaceProject[] {
-    return [];
-  }
-
-  async createWorkspace(_request: ProviderWorkspaceRequest): Promise<ProviderWorkspaceResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider cannot create workspaces.",
-    };
-  }
-
-  async spawnWorkspaceAgent(
-    _request: ProviderWorkspaceAgentRequest,
-  ): Promise<ProviderWorkspaceResult> {
-    return { status: ACT_RESULT_STATUS.UNSUPPORTED, reason: "This provider cannot add agents." };
-  }
-
-  async renameWorkspace(_request: ProviderWorkspaceRenameRequest): Promise<ProviderActResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider cannot rename workspaces.",
-    };
-  }
-
-  async renameSession(_request: ProviderSessionRenameRequest): Promise<ProviderActResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider cannot rename sessions.",
-    };
-  }
-
-  async readTranscript(_providerSessionId: string): Promise<ProviderTranscriptResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider keeps no transcript this build can read.",
-    };
-  }
-
-  async readTranscriptSince(
-    _providerSessionId: string,
-    _cursor?: string,
-  ): Promise<ProviderTranscriptSinceResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider keeps no transcript this build can read.",
-    };
-  }
-
-  async readConversation(
-    _request: ProviderConversationRequest,
-  ): Promise<ProviderConversationResult> {
-    return {
-      status: ACT_RESULT_STATUS.UNSUPPORTED,
-      reason: "This provider documents no conversation read this build carries.",
-    };
-  }
-}
-
-/**
- * Whether a CLI-observed provider can be observed right now, without ever
- * exposing the login behind it — the CLI analogue of the credential sources a keyed provider reports.
- * Unknown is the state before the first pass has asked; the other three are
- * what the latest pass learned from the provider's own CLI.
- */
-export const CLI_CONNECTION = {
-  UNKNOWN: "unknown",
-  CONNECTED: "connected",
-  SIGNED_OUT: "signed-out",
-  CLI_MISSING: "cli-missing",
-} as const;
-
-export type CliConnection = (typeof CLI_CONNECTION)[keyof typeof CLI_CONNECTION];
