+++ packages/session/src/workspace-agents.ts
@@ -3,10 +3,73 @@ import {
   isProviderId,
   PROVIDER_ID,
   type ProviderId,
-  type WorkspaceAgentKindSelection,
-  type WorkspaceAgentModels,
-  type WorkspaceAgentSelection,
-} from "./providers.js";
+  type SUPERSET_WORKSPACE_PROVIDER_ID,
+} from "./provider-identity.js";
+
+/**
+ * One model an agent runs: the id its provider's endpoints take, and the name
+ * a person reads. The label stands alone — no vendor beside it — because a
+ * model's own name already says whose it is.
+ */
+export interface WorkspaceAgentModel {
+  id: string;
+  label: string;
+}
+
+/**
+ * One agent kind a provider's creation endpoints take, with the models it
+ * runs and the effort levels it thinks at — all exactly as the provider
+ * documents them. An app declares these as build-fixed tables, the way it
+ * fixes any other documented value set, so a choice is only ever offered from
+ * what the build knows the provider takes. An agent that takes no effort
+ * levels documents an empty list, and is simply never offered one.
+ */
+export interface WorkspaceAgentModels {
+  agent: string;
+  models: readonly WorkspaceAgentModel[];
+  efforts: readonly string[];
+}
+
+/**
+ * The agent kind, model, and optionally effort a user chose for new
+ * workspaces, as one value on purpose: a model id or an effort level only
+ * means anything beside the agent that runs it, and fields stored apart could
+ * recombine into a pairing no table ever listed. Effort is optional inside
+ * the pair — absent, the provider's own default effort stands.
+ */
+export interface WorkspaceAgentSelection {
+  /** The agent kind, exactly as the provider's documented set names it. */
+  agent: string;
+  /** The model id, exactly as documented for that agent. */
+  model: string;
+  /** The effort level, exactly as documented for that agent, when chosen. */
+  effort?: string;
+}
+
+/**
+ * The agent kind a user chose for new workspaces on a provider whose agent
+ * kinds are observed presets rather than a build-fixed table, and which
+ * documents no model choice (Superset today). The `never` fields are the
+ * boundary stated as a type: no model or effort can ride beside a kind the
+ * provider's endpoints take alone.
+ */
+export interface WorkspaceAgentKindSelection {
+  agent: string;
+  model?: never;
+  effort?: never;
+}
+
+/**
+ * The chosen defaults for new workspaces, one entry per workspace provider
+ * that documents an agent choice at all: a provider with a build-fixed models
+ * table holds a full selection, and Superset holds the kind alone. Local
+ * Conductor's creation link documents no agent, model, or name, so it can
+ * hold no entry.
+ */
+export type WorkspaceAgentDefaults = Readonly<
+  Partial<Record<ProviderId, WorkspaceAgentSelection>> &
+    Partial<Record<typeof SUPERSET_WORKSPACE_PROVIDER_ID, WorkspaceAgentKindSelection>>
+>;
 
 /**
  * The agent kinds, models, and effort levels each provider's creation
